backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
Pillow==10.4.0
aiofiles==24.1.0
